Extension works without content_scripts in manifest
These changes reduce the necessary permissions needed to run this
extension by only giving the scripts access to the current tab at
runtime. This uses the activeTab permission instead of setting the
context_scripts in the manifest and does not require any permissions
from the user when they install the addon.